moved IO functions to lib/utils.ts
